Made log4j dependency explicit in subprojects
Previously this dependency was defined in the parent project (master build),
which caused it to be included in all subprojects.

When I will create an EAR module, it needn't have the log4j jar in it, so I am
moving the log4j dependency to the subprojects that require it.